:bug: Correct player rotation over network
Fixes an issue where all instances of the player class were rotating to
match the clients mouse movement thus overwriting the servers rotation
instructions. Includes the player rotation as part of the state
update packet, this reduces the number of packets sent and ensures that
the rotation is synced correctly across clients.